+++ packages/dashboard-viewer/src/views/Dashboard.vue
@@ -5,6 +5,7 @@
       flat
       v-if="!isTV && !isFullscreen"
       :color="$vuetify.theme.dark ? '#121212' : 'grey lighten-5'"
+      style="z-index: 99"
     >
       <img
         :src="require(`@shopworx/assets/logo/${shopworxLogo}.png`)"
@@ -20,17 +21,20 @@
         <portal-target name="dashboard-title"></portal-target>
       </v-toolbar-title>
       <v-spacer></v-spacer>
-      <send-to-tv />
-      <v-btn
-        class="text-none mr-2"
-        color="primary"
-        outlined
-        small
-        @click="previewDashboard"
-      >
-        <v-icon left>mdi-television</v-icon>
-        Preview
-      </v-btn>
+      <portal-target name="dashboard-actions" />
+      <template v-if="showHeaderButtons">
+        <send-to-tv />
+        <v-btn
+          class="text-none mr-2"
+          color="primary"
+          outlined
+          small
+          @click="previewDashboard"
+        >
+          <v-icon left>mdi-television</v-icon>
+          Preview
+        </v-btn>
+      </template>
       <v-btn
         class="text-none mr-2"
         color="primary"
@@ -72,7 +76,7 @@ export default {
   },
   computed: {
     ...mapGetters('helper', ['isTV']),
-    ...mapState('helper', ['isFullscreen']),
+    ...mapState('helper', ['isFullscreen', 'showHeaderButtons']),
     shopworxLogo() {
       return this.$vuetify.theme.dark
         ? 'shopworx-dark'