+++ packages/dashboards/utility-dashboard/src/utility-dashboard-routes.js
@@ -1,0 +1,18 @@
+const routes = [
+  {
+    path: 'utility',
+    component: () => import(/* webpackChunkName: "utility-dashboard" */ './views/Index.vue'),
+    meta: {
+      permissionRequired: true,
+    },
+    children: [
+      {
+        path: '',
+        name: 'utilityDashboard',
+        component: () => import(/* webpackChunkName: "utility-dashboard" */ './views/UtilityDashboard.vue'),
+      },
+    ],
+  },
+];
+
+export default routes;